+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 # Changelog
 
+## indev 0.1.1 (12/8/2020)
+- Setting up Express Routes
+- Applying EJS as render engine
+- Setting up placeholder pages and partials
+
+## indev 0.1.0 (12/8/2020)
+- Database Connection with MongoDB Atlas
+- Express setup
+- NPM Packages Set Up
+
 ## Initial Commit (12/8/2020)
 - Setting Up the Repository